add filter to starboard to prevent unnecessary code from running
Check if [the emoji added to a message] is the starboard emoji or the downvote cross mark. Else, end the event function.
Also add a url for the people who submit a QOTW or someone whose message gets starboarded, in which the url is their ID
Add convert_unit function to convert imperial/metric into each other