Rewrite MissingAttributesCheckFailure to avoid *args.
You can raise a new CheckFailure with a module and a list of attributes. The first argument is a module, and the extra attribute arguments are wrapped into one argument. However, if you don't give a Module name, the first attribute argument is interpreted as the module name. As a result, the list of arguments may be empty. This commit aims to resolve such scenarios by making both arguments required.
Another method would be to check if the first argument (module) is a valid module name, but that could be prone to error if attributes and modules have the same name. So I decided to just do it this way.