Trying to salvage Wolfram Alpha by adding typed dicts.
Why does the "success" variable not always exist.... it's a true/false variable.. Who thought it was a good idea not to always return a "success=False" when no response can be given? That's dumb. It's not mentioned anywhere in the docs either. -_- Welp, maybe this more-expansive TypedDict chaos helps catch these odd-one-outs somewhat.